+++ include/trix/trix_version.h
@@ -5,8 +5,8 @@
 
 #define TRIX_VERSION_MAJOR 1
 #define TRIX_VERSION_MINOR 1
-#define TRIX_VERSION_PATCH 0
-#define TRIX_VERSION_STRING "1.1.0"
+#define TRIX_VERSION_PATCH 1
+#define TRIX_VERSION_STRING "1.1.1"
 
 /* Single integer: 10000*major + 100*minor + patch */
 #define TRIX_VERSION_NUMBER \